dash/requirements.txt
attrs==23.2.0
aws-cdk-lib>=2.152.0
cattrs==23.2.3
constructs>=10.0.0,<11.0.0
importlib_resources==6.4.2
jsii==1.102.0
publication==0.0.3
python-dateutil==2.9.0.post0
six==1.16.0
typeguard==2.13.3
typing_extensions==4.12.2
